requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.9.0=py36_0
argon2-cffi=20.1.0=py36h8c4c3a4_1
attrs=19.3.0=py_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.1=py_0
blas=1.0=mkl
bleach=3.1.5=pyh9f0ad1d_0
blinker=1.4=py_1
boost-cpp=1.72.0=h7b93d67_2
brotlipy=0.7.0=py36h8c4c3a4_1000
bzip2=1.0.8=h516909a_2
c-ares=1.16.1=h516909a_0
ca-certificates=2020.6.20=hecda079_0
cachetools=4.1.1=py_0
cairo=1.16.0=h3fc0475_1005
certifi=2020.6.20=py36h9f0ad1d_0
cffi=1.11.5=py36_0
cfitsio=3.470=hce51eda_6
chardet=3.0.4=py36h9f0ad1d_1006
click=7.1.2=pyh9f0ad1d_0
click-plugins=1.1.1=py_0
cligj=0.5.0=py_0
cryptography=3.0=py36h45558ae_0
cudatoolkit=9.2=0
curl=7.71.1=he644dc0_5
cycler=0.10.0=py_2
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
entrypoints=0.3=py36h9f0ad1d_1001
expat=2.2.9=he1b5a44_2
ffmpeg=4.3.1=h167e202_0
fiona=1.8.13=py36h9956732_2
fontconfig=2.13.1=h1056068_1002
freetype=2.10.2=h5ab3b9f_0
freexl=1.0.5=h516909a_1002
gdal=3.1.2=py36hcf62b00_1
geopandas=0.8.1=py_0
geos=3.8.1=he1b5a44_0
geotiff=1.6.0=ha04d9d0_1
giflib=5.2.1=h516909a_2
glib=2.65.0=h3eb4bd4_0
gmp=6.2.0=he1b5a44_2
gnutls=3.6.13=h79a8f9a_0
google-auth=1.20.1=py_0
google-auth-oauthlib=0.4.1=py_2
grpcio=1.31.0=py36h769ab6c_0
h5py=2.10.0=py36hd6299e0_1
hdf4=4.2.13=hf30be14_1003
hdf5=1.10.6=nompi_h3c11f04_101
icu=67.1=he1b5a44_0
idna=2.10=pyh9f0ad1d_0
imageio=2.9.0=py_0
imageio-ffmpeg=0.4.2=py_0
importlib-metadata=1.7.0=py36h9f0ad1d_0
importlib_metadata=1.7.0=0
intel-openmp=2020.1=217
ipykernel=5.3.4=py36h95af2a2_0
ipython=5.8.0=py36_1
ipython_genutils=0.2.0=py_1
jinja2=2.11.2=pyh9f0ad1d_0
joblib=0.16.0=py_0
jpeg=9d=h516909a_0
json-c=0.13.1=hbfbb72e_1002
json5=0.9.4=pyh9f0ad1d_0
jsonschema=3.2.0=py36h9f0ad1d_1
jupyter_client=6.1.6=py_0
jupyter_core=4.6.3=py36h9f0ad1d_1
jupyterlab=2.2.4=py_0
jupyterlab_server=1.2.0=py_0
kealib=1.4.13=h33137a7_1
kiwisolver=1.2.0=py36hdb11119_0
krb5=1.17.1=hfafb76e_2
lame=3.100=h14c3975_1001
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libcurl=7.71.1=hcdd3856_5
libdap4=3.20.6=h1d1bd15_1
libedit=3.1.20191231=h14c3975_1
libev=4.33=h516909a_0
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgdal=3.1.2=hb2a6f5f_1
libgfortran-ng=7.3.0=hdf63c60_0
libiconv=1.16=h516909a_0
libkml=1.3.0=hb574062_1011
libnetcdf=4.7.4=nompi_h84807e1_105
libnghttp2=1.41.0=hab1572f_1
libpng=1.6.37=hbc83047_0
libpq=12.3=h5513abc_0
libprotobuf=3.12.4=h8b12597_0
libsodium=1.0.18=h516909a_0
libspatialindex=1.9.3=he1b5a44_3
libspatialite=4.3.0a=h57f1b35_1039
libssh2=1.9.0=hab1572f_5
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=2.32.1=h14c3975_1000
libwebp-base=1.1.0=h516909a_3
libxcb=1.13=h14c3975_1002
libxml2=2.9.10=h68273f3_2
lz4-c=1.9.2=he6710b0_1
markdown=3.2.2=py_0
markupsafe=1.1.1=py36h8c4c3a4_1
matplotlib=3.3.1=0
matplotlib-base=3.3.1=py36h5ffbc53_0
mistune=0.8.4=py36h8c4c3a4_1001
mkl=2020.1=217
mkl-service=2.3.0=py36he904b0f_0
mkl_fft=1.1.0=py36h23d657b_0
mkl_random=1.1.1=py36h0573a6f_0
moviepy=1.0.1=py_0
munch=2.5.0=py_0
nbconvert=5.6.1=py36h9f0ad1d_1
nbformat=5.0.7=py_0
ncurses=6.2=he6710b0_1
nettle=3.4.1=h1bed415_1002
ninja=1.10.0=py36hfd86e86_0
notebook=6.1.3=py36h9f0ad1d_0
numpy=1.19.1=py36hbc911f0_0
numpy-base=1.19.1=py36hfa32c7d_0
oauthlib=3.0.1=py_0
olefile=0.46=py36_0
openh264=2.1.1=h8b12597_0
openjpeg=2.3.1=h981e76c_3
openssl=1.1.1g=h516909a_1
packaging=20.4=pyh9f0ad1d_0
pandas=1.1.0=py36he6710b0_0
pandoc=2.10.1=h516909a_0
pandocfilters=1.4.2=py_1
pcre=8.44=he1b5a44_0
pexpect=4.8.0=py36h9f0ad1d_1
pickleshare=0.7.5=py36h9f0ad1d_1001
pillow=7.2.0=py36hb39fc2d_0
pip=20.2.2=py36_0
pixman=0.38.0=h516909a_1003
poppler=0.89.0=h4190859_1
poppler-data=0.4.9=1
postgresql=12.3=h8573dbc_0
proglog=0.1.9=py_0
proj=7.1.0=h966b41f_1
prometheus_client=0.8.0=pyh9f0ad1d_0
prompt_toolkit=1.0.15=py_1
protobuf=3.12.4=py36h831f99a_0
psycopg2=2.8.5=py36h1ba5d50_0
pthread-stubs=0.4=h14c3975_1001
ptyprocess=0.6.0=py_1001
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.20=pyh9f0ad1d_2
pygments=2.6.1=py_0
pyjwt=1.7.1=py_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=pyh9f0ad1d_0
pyproj=2.6.1.post1=py36hf866a91_1
pyrsistent=0.16.0=py36h8c4c3a4_0
pysocks=1.7.1=py36h9f0ad1d_1
python=3.6.10=h7579374_2
python-dateutil=2.8.1=py_0
python_abi=3.6=1_cp36m
pytorch=1.6.0=py3.6_cuda9.2.148_cudnn7.6.3_0
pytz=2020.1=py_0
pyzmq=19.0.2=py36h9947dbf_0
readline=8.0=h7b6447c_0
requests=2.24.0=pyh9f0ad1d_0
requests-oauthlib=1.3.0=pyh9f0ad1d_0
rsa=4.6=pyh9f0ad1d_0
rtree=0.9.4=py36he053a7a_1
scikit-learn=0.23.1=py36h423224d_0
scipy=1.5.0=py36h0b6359f_0
seaborn=0.10.1=py_0
send2trash=1.5.0=py_0
setuptools=49.6.0=py36_0
shapely=1.7.0=py36h3d6ee9d_3
simplegeneric=0.8.1=py_1
six=1.15.0=py_0
sqlite=3.32.3=h62c20be_0
tbb=2020.1=hc9558a2_0
tensorboard=2.3.0=py_0
tensorboard-plugin-wit=1.6.0=pyh9f0ad1d_0
terminado=0.8.3=py36h9f0ad1d_1
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tiledb=2.0.8=h3effe38_0
tk=8.6.10=hbc83047_0
torchsummary=1.5.1=pypi_0
torchvision=0.7.0=py36_cu92
tornado=6.0.4=py36h8c4c3a4_1
tqdm=4.48.2=pyh9f0ad1d_0
traitlets=4.3.3=py36h9f0ad1d_1
tzcode=2020a=h516909a_0
urllib3=1.25.10=py_0
wcwidth=0.2.5=pyh9f0ad1d_1
webencodings=0.5.1=py_1
werkzeug=1.0.1=pyh9f0ad1d_0
wheel=0.34.2=py36_0
x264=1!152.20180806=h14c3975_0
xerces-c=3.2.3=hfe33f54_1
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libice=1.0.10=h516909a_0
xorg-libsm=1.2.3=h84519dc_1000
xorg-libx11=1.6.11=h516909a_0
xorg-libxau=1.0.9=h14c3975_0
xorg-libxdmcp=1.1.3=h516909a_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxrender=0.9.10=h516909a_1002
xorg-renderproto=0.11.1=h14c3975_1002
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.5=h7b6447c_0
zeromq=4.3.2=he1b5a44_3
zipp=3.1.0=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
